Remove the casts, they are no longer needed
git-svn-id: svn://svn.cc65.org/cc65/trunk@1478 b7a2c559-68d2-44c3-8de9-860c34a00d81

+++ testcode/lib/div-test.c
@@ -13,7 +13,7 @@
 static bool test(int dividend, int divisor) {
 	div_t result;
 
-	(long)result = (long)div(dividend, divisor);
+       	result = div(dividend, divisor);
 	printf("%+d/%+d= %+d, %+d%%%+d= %+d, div()= %+d, %+d\n",
 		dividend, divisor, dividend / divisor,
 		dividend, divisor, dividend % divisor,
